garment with token

+++ lib/repository/garment_repo.dart
@@ -8,29 +8,32 @@ import 'package:wardrobe_mobile/repository/APIConstant.dart';
 class GarmentRepository{
   Future<bool> createGarment(GarmentModel shirt) async {
       try{
-        var url = Uri.parse(APIConstant.addGarmentURL);
-        print(url);
-        var body = json.encode({
-          "name":shirt.name,
-          "brand": shirt.brand,
-          "country": shirt.country,
-          "size": shirt.size,
-          "colour": shirt.colour,
-          "colour_name" : shirt.colour_name,
-          "status": true,
-          "image": await _convertImageToBase64(shirt.image!),
-        });
-        print(body);
-        var header = {
-          "Content-Type": "application/json",
-        };
-        var response = await http.post(url, headers: header, body: body);
-        print(response.statusCode);
-        if (response.statusCode == 201) {
-          return true;
-        } else {
-          return false;
+        var pref = await SharedPreferences.getInstance();
+        String? token = pref.getString('token') ;
+        if (token != null){
+          var url = Uri.parse(APIConstant.addGarmentURL);
+          var body = json.encode({
+            "name":shirt.name,
+            "brand": shirt.brand,
+            "country": shirt.country,
+            "size": shirt.size,
+            "colour": shirt.colour,
+            "colour_name" : shirt.colour_name,
+            "status": true,
+            "image": await _convertImageToBase64(shirt.image!),
+          });
+          print(body);
+          var header = {
+            "Content-Type": "application/json",
+            "Authorization": "Bearer $token"
+          };
+          var response = await http.post(url, headers: header, body: body);
+          print(response.statusCode);
+          if (response.statusCode == 201) {
+            return true;
+          }
         }
+        return false;
       } catch (e) {
         return false;
       }
@@ -96,39 +99,49 @@ class GarmentRepository{
 
   Future<GarmentModel?> getOneGarment(String id) async{
     try{
-      var url = Uri.parse(APIConstant.getOneGarmentURL + id);
-      var header = {
-        "Content-Type": "application/json",
-      };
-      var response = await http.get(url, headers: header);
-      if (response.statusCode == 200){
-                print(jsonDecode(response.body));
+      var pref = await SharedPreferences.getInstance();
+      String? token = pref.getString('token') ;
+      if (token != null){
+        var url = Uri.parse(APIConstant.getOneGarmentURL + id);
+        var header = {
+          "Content-Type": "application/json",
+          "Authorization": "Bearer $token"
+        };
+        var response = await http.get(url, headers: header);
+        if (response.statusCode == 200){
+          print(jsonDecode(response.body));
 
-        var result = jsonDecode(response.body)['garments'];
-        return GarmentModel.fromJson(result);
+          var result = jsonDecode(response.body)['garments'];
+          return GarmentModel.fromJson(result);
+        }
       }
-
+      return null;
     } catch (e){
       print(e.toString());
       return null;
     }
-    return null;
   }
 
   Future<bool> deleteGarment(String id) async{
     try{
-      var url = Uri.parse(APIConstant.deleteGarmentURL);
-      print(id);
-      var body = json.encode({
-        "id" : id,
-      });
-      var header = {
-        "Content-Type": "application/json",
-      };
-      var response = await http.delete(url, headers: header, body:body);
+      var pref = await SharedPreferences.getInstance();
+      String? token = pref.getString('token') ;
+      if (token != null){
 
-      if (response.statusCode == 200){
-        return true;
+        var url = Uri.parse(APIConstant.deleteGarmentURL);
+        print(id);
+        var body = json.encode({
+          "id" : id,
+        });
+        var header = {
+          "Content-Type": "application/json",
+          "Authorization": "Bearer $token"
+        };
+        var response = await http.delete(url, headers: header, body:body);
+
+        if (response.statusCode == 200){
+          return true;
+        }
       }
       return false;
     } catch(e){
@@ -140,28 +153,29 @@ class GarmentRepository{
 
   Future<bool> updateGarment(GarmentModel gmt) async {
     try{
-      var url = Uri.parse(APIConstant.updateGarmentURL + gmt.id!);
-      print(url);
-      var header = {
-        "Content-Type": "application/json"
-      };
-      var body = json.encode({
-        "brand": gmt.brand,
-        "name": gmt.name,
-        "colour": gmt.colour,
-        "country": gmt.country,
-        "size" : gmt.size,
-        "status": true,
-      });
+      var pref = await SharedPreferences.getInstance();
+      String? token = pref.getString('token') ;
+      if (token != null){
+        var url = Uri.parse(APIConstant.updateGarmentURL + gmt.id!);
+        var header = {
+          "Content-Type": "application/json",
+          "Authorization": "Bearer $token"
+        };
+        var body = json.encode({
+          "brand": gmt.brand,
+          "name": gmt.name,
+          "colour": gmt.colour,
+          "country": gmt.country,
+          "size" : gmt.size,
+          "status": true,
+        });
 
-      var response = await http.put(url, headers: header, body: body);
-      if (response.statusCode == 200){
-        return true;
+        var response = await http.put(url, headers: header, body: body);
+        if (response.statusCode == 200){
+          return true;
+        }
       }
-      else {
-        return false;
-      }
-
+      return false;
     } catch (e){
       return false;
     }
